# delivery/api-dog/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/checkpoints": {
      "post": {
        "operationId": "CheckpointsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Checkpoints"
        ]
      },
      "get": {
        "description": "**Access:** `public` — optional `status` filters the registry.",
        "operationId": "CheckpointsController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "OPEN",
                "CLOSED",
                "UNKNOWN"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List checkpoints (paginated, optional status filter)",
        "tags": [
          "Checkpoints"
        ]
      }
    },
...
    "/api/v1/incidents": {
      "post": {
        "description": "**Access:** `moderator` or `admin` (JWT). Course-aligned: only authorized staff create official incidents.",
        "operationId": "IncidentsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create incident",
        "tags": [
          "Incidents"
        ]
      },
      "get": {
        "description": "**Access:** `public` — filter by `status`, `severity`, and case-insensitive `type`; sort `createdAt` with `sort=asc|desc`.",
        "operationId": "IncidentsController_findAll",
        "parameters": [
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Incident type string (case-insensitive match)",
            "schema": {
              "example": "closure",
              "type": "string"
            }
          },
          {
            "name": "severity",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "LOW",
                "MEDIUM",
                "HIGH"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "OPEN",
                "VERIFIED",
                "CLOSED"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List incidents (paginated, filterable, sortable)",
        "tags": [
          "Incidents"
        ]
      }
    },
    "/api/v1/incidents/{id}": {
      "get": {
        "description": "**Access:** `public`",
        "operationId": "IncidentsController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get incident by id",
        "tags": [
          "Incidents"
        ]
      },
      "put": {
        "description": "**Access:** `moderator` or `admin`",
        "operationId": "IncidentsController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update incident",
        "tags": [
          "Incidents"
        ]
      },
      "delete": {
        "description": "**Access:** `moderator` or `admin`",
        "operationId": "IncidentsController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete incident",
        "tags": [
          "Incidents"
        ]
      }
    },
    "/api/v1/incidents/{id}/verify": {
      "patch": {
        "description": "**Access:** `moderator` or `admin` — sets status to VERIFIED and triggers subscriber alerts.",
        "operationId": "IncidentsController_verify",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify incident",
        "tags": [
          "Incidents"
        ]
      }
    },
    "/api/v1/incidents/{id}/close": {
      "patch": {
        "description": "**Access:** `moderator` or `admin`",
        "operationId": "IncidentsController_close",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Close incident",
        "tags": [
          "Incidents"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wasel Palestine - API Reference",
    "description": "REST API for Wasel Palestine.\n\nImport this file into API Dog:\n- File: `delivery/api-dog/openapi.json` (this export)\n- Or live: `GET /openapi.json` from a running server",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}